Add separate v4l2 camera gst pipeline to avoid issues with default gstCamera issues
With the latest changes to jetson-utils involving updates to the
constructed gstreamer pipelines, v4l2 (usb) cameras cannot be properly
accessed from within an application that attempts to capture frames from
them. Therefore, a proper pipeline was added for the case of devices
whose uri is of type v4l2.